voicebot_tutor/requirements.txt
openai==0.27.0
SpeechRecognition==3.8.1
pyaudio==0.2.11
pygame==2.1.2
python-dotenv==0.19.2
numpy==1.21.5